Fix PR remarks
