+++ RiotSwiftUI/Modules/Authentication/ChoosePassword/Test/UI/AuthenticationChoosePasswordUITests.swift
@@ -45,64 +45,74 @@ class AuthenticationChoosePasswordUITests: MockScreenTest {
         XCTAssertTrue(app.staticTexts["titleLabel"].exists, "The title should be shown.")
         XCTAssertTrue(app.staticTexts["messageLabel"].exists, "The message should be shown.")
         
-        let passwordTextField = app.textFields["passwordTextField"]
+        let passwordTextField = app.secureTextFields["passwordTextField"]
         XCTAssertTrue(passwordTextField.exists, "The text field should be shown.")
-        XCTAssertEqual(passwordTextField.value as? String, "New Password", "The text field should be showing the placeholder before text is input.")
+        XCTAssertEqual(passwordTextField.label, "New Password", "The text field should be showing the placeholder before text is input.")
         
         let submitButton = app.buttons["submitButton"]
         XCTAssertTrue(submitButton.exists, "The submit button should be shown.")
         XCTAssertFalse(submitButton.isEnabled, "The submit button should be disabled before text is input.")
 
-        let doNotSignoutAllDevicesImage = app.images["doNotSignoutAllDevices"]
-        XCTAssertTrue(doNotSignoutAllDevicesImage.exists, "Sign out all devices should be unchecked")
+        let signoutAllDevicesToggle = app.switches["signoutAllDevicesToggle"]
+        XCTAssertTrue(signoutAllDevicesToggle.exists, "Sign out all devices toggle should exist")
+        XCTAssertFalse(signoutAllDevicesToggle.isOn, "Sign out all devices should be unchecked")
     }
     
     func verifyEnteredInvalidPassword() {
         XCTAssertTrue(app.staticTexts["titleLabel"].exists, "The title should be shown.")
         XCTAssertTrue(app.staticTexts["messageLabel"].exists, "The message should be shown.")
 
-        let passwordTextField = app.textFields["passwordTextField"]
+        let passwordTextField = app.secureTextFields["passwordTextField"]
         XCTAssertTrue(passwordTextField.exists, "The text field should be shown.")
-        XCTAssertEqual(passwordTextField.value as? String, "1234", "The text field should be showing the placeholder before text is input.")
+        XCTAssertEqual(passwordTextField.value as? String, "••••", "The text field should be showing the placeholder before text is input.")
 
         let submitButton = app.buttons["submitButton"]
         XCTAssertTrue(submitButton.exists, "The submit button should be shown.")
         XCTAssertFalse(submitButton.isEnabled, "The submit button should be disabled when password is invalid.")
 
-        let doNotSignoutAllDevicesImage = app.images["doNotSignoutAllDevices"]
-        XCTAssertTrue(doNotSignoutAllDevicesImage.exists, "Sign out all devices should be unchecked")
+        let signoutAllDevicesToggle = app.switches["signoutAllDevicesToggle"]
+        XCTAssertTrue(signoutAllDevicesToggle.exists, "Sign out all devices toggle should exist")
+        XCTAssertFalse(signoutAllDevicesToggle.isOn, "Sign out all devices should be unchecked")
     }
     
     func verifyEnteredValidPassword() {
         XCTAssertTrue(app.staticTexts["titleLabel"].exists, "The title should be shown.")
         XCTAssertTrue(app.staticTexts["messageLabel"].exists, "The message should be shown.")
 
-        let passwordTextField = app.textFields["passwordTextField"]
+        let passwordTextField = app.secureTextFields["passwordTextField"]
         XCTAssertTrue(passwordTextField.exists, "The text field should be shown.")
-        XCTAssertEqual(passwordTextField.value as? String, "12345678", "The text field should be showing the placeholder before text is input.")
+        XCTAssertEqual(passwordTextField.value as? String, "••••••••", "The text field should be showing the placeholder before text is input.")
 
         let submitButton = app.buttons["submitButton"]
         XCTAssertTrue(submitButton.exists, "The submit button should be shown.")
         XCTAssertTrue(submitButton.isEnabled, "The submit button should be enabled after password is valid.")
 
-        let doNotSignoutAllDevicesImage = app.images["doNotSignoutAllDevices"]
-        XCTAssertTrue(doNotSignoutAllDevicesImage.exists, "Sign out all devices should be unchecked")
+        let signoutAllDevicesToggle = app.switches["signoutAllDevicesToggle"]
+        XCTAssertTrue(signoutAllDevicesToggle.exists, "Sign out all devices toggle should exist")
+        XCTAssertFalse(signoutAllDevicesToggle.isOn, "Sign out all devices should be unchecked")
     }
 
     func verifyEnteredValidPasswordAndSignoutAllDevicesChecked() {
         XCTAssertTrue(app.staticTexts["titleLabel"].exists, "The title should be shown.")
         XCTAssertTrue(app.staticTexts["messageLabel"].exists, "The message should be shown.")
 
-        let passwordTextField = app.textFields["passwordTextField"]
+        let passwordTextField = app.secureTextFields["passwordTextField"]
         XCTAssertTrue(passwordTextField.exists, "The text field should be shown.")
-        XCTAssertEqual(passwordTextField.value as? String, "12345678", "The text field should be showing the placeholder before text is input.")
+        XCTAssertEqual(passwordTextField.value as? String, "••••••••", "The text field should be showing the placeholder before text is input.")
 
         let submitButton = app.buttons["submitButton"]
         XCTAssertTrue(submitButton.exists, "The submit button should be shown.")
         XCTAssertTrue(submitButton.isEnabled, "The submit button should be enabled after password is valid.")
 
-        let signoutAllDevicesImage = app.images["signoutAllDevices"]
-        XCTAssertTrue(signoutAllDevicesImage.exists, "Sign out all devices should be checked")
+        let signoutAllDevicesToggle = app.switches["signoutAllDevicesToggle"]
+        XCTAssertTrue(signoutAllDevicesToggle.exists, "Sign out all devices toggle should exist")
+        XCTAssertTrue(signoutAllDevicesToggle.isOn, "Sign out all devices should be checked")
     }
 
+}
+
+extension XCUIElement {
+    var isOn: Bool {
+        (value as? String) == "1"
+    }
 }
